# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bitwarden client API contract (Cloudwarden)
  version: 0.1.0
  description: |-
    Wire contract for the Bitwarden client API as called by the official Bitwarden clients (web vault, browser extension, desktop, CLI). It exists so Cloudwarden can be built against a written contract instead of any server implementation.

    **Source of truth.** Every operation records the clients repository release tag it was read from in `x-source-version` and the file in `x-source-file`. The primary tag is `web-v2026.9.1` of github.com/bitwarden/clients. A few calls moved into the Bitwarden SDK (a separate repository) after 2025; those operations cite the older tag where the client still contained the call (`web-v2025.8.0` or `web-v2025.1.0`). Shapes that exist only inside the SDK (the `send_access` grant, enum values for policy types) are flagged in their descriptions.

    **Services and paths.** One base URL serves several services. Official clients derive: API `<base>/api`, identity `<base>/identity`, notifications `<base>/notifications`, icons `<base>/icons`, events `<base>/events`. Paths in this document include that prefix.

...

...

    **Errors.** API errors use `ErrorResponseModel` (`message`, `validationErrors`, `object: error`). Identity errors use OAuth fields plus `ErrorModel`. HTTP 429 is shown as a rate limit message. Successful empty responses may be 200 or 204; clients treat both as success.

...
  license:
    name: AGPL-3.0-only
...
  contact:
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /identity/accounts/prelogin:
    post:
      tags:
      - Identity
      operationId: preloginLegacy
      summary: Get KDF settings (legacy)
      description: Flat-field prelogin used by clients up to 2025.8. Still called by older mobile and CLI releases.
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreloginRequest'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreloginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security: []
      x-source-version: web-v2025.8.0
      x-source-file: libs/common/src/services/api.service.ts
  /identity/accounts/prelogin/password:
    post:
      tags:
      - Identity
      operationId: preloginPassword
      summary: Get KDF settings and salt for an email
      description: 'Current prelogin call. Must not reveal whether the account exists: unknown emails should receive
        default KDF settings and a deterministic salt.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordPreloginRequest'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordPreloginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security: []
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/auth/password-prelogin/password-prelogin-api.service.ts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/identity/connect/token`.
    sendAccessToken:
      type: http
      scheme: bearer
      description: Send access token from the `send_access` grant (not the user token).
    hubAccessToken:
      type: apiKey
      in: query
      name: access_token
      description: User access token passed as a query parameter on the WebSocket upgrade.
  responses:
    BadRequest:
      description: Validation or business rule failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseModel'
...
    TooManyRequests:
      description: Rate limited. Clients show a generic rate limit message.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseModel'
  schemas:
...
    ErrorResponseModel:
      type: object
      description: Error body returned by the API service (`/api/*`, `/events/*`). The client reads `Message` and
        `ValidationErrors` case-insensitively; a 429 status is shown as a rate-limit message regardless of body.
        `application/problem+json` bodies (RFC 9457) are also accepted by the client for some endpoints.
      properties:
        message:
          type:
          - string
          - 'null'
        validationErrors:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field path to messages. Keys with a bracket index (for example `Ciphers[0].Name`)
            are shown by clients with the item prefix.
        object:
          type: string
          enum:
          - error
...
    KdfConfig:
      type: object
      description: KDF parameters. `memory` (MiB) and `parallelism` are required for Argon2id and omitted for PBKDF2.
      properties:
        kdfType:
          $ref: '#/components/schemas/KdfType'
        iterations:
          type: integer
        memory:
          type:
          - integer
          - 'null'
        parallelism:
          type:
          - integer
          - 'null'
      required:
      - kdfType
      - iterations
...
    KdfType:
      type: integer
      description: Key derivation function.
      enum:
      - 0
      - 1
      x-enum-varnames:
      - PBKDF2_SHA256
      - Argon2id
...
    PasswordPreloginRequest:
      type: object
      properties:
        email:
          type: string
      required:
      - email
    PasswordPreloginResponse:
      type: object
      description: Current prelogin response. The client throws if `kdfType` or `iterations` are missing or not
        numbers.
      properties:
        kdfSettings:
          $ref: '#/components/schemas/KdfConfig'
        salt:
          type: string
      required:
      - kdfSettings
      - salt
...
    PreloginRequest:
      type: object
      properties:
        email:
          type: string
      required:
      - email
    PreloginResponse:
      type: object
      description: Legacy prelogin response (flat KDF fields).
      properties:
        kdf:
          $ref: '#/components/schemas/KdfType'
        kdfIterations:
          type: integer
        kdfMemory:
          type:
          - integer
          - 'null'
        kdfParallelism:
          type:
          - integer
          - 'null'
      required:
      - kdf
      - kdfIterations
...